refactor: rename useImageUpload.ts to useFileUpload.ts
